コードレビュー指摘対応: 自動番号リセット修正・フォントサイズ精度改善・色変換順序修正
- 箇条書き指定なし段落(PPrなし・BuNone含む)で自動番号カウンタをリセットするよう修正
- FontStyle.Sizeをintからfloat64に変更し半ポイントサイズ(11.5pt等)を正確に表現
- loadRelsTypedにファイル不在時の挙動をドキュメント追加
- 色変換(tint/shade/lumMod/lumOff)をXML出現順に適用するようOOXML仕様に準拠

+++ internal/pptx/types.go
@@ -80,8 +80,8 @@ type RichTextRun struct {
 
 // FontStyle はフォント情報
 type FontStyle struct {
-	Name          string `json:"name,omitempty"`
-	Size          int    `json:"size,omitempty"`
+	Name          string  `json:"name,omitempty"`
+	Size          float64 `json:"size,omitempty"`
 	Bold          bool   `json:"bold,omitempty"`
 	Italic        bool   `json:"italic,omitempty"`
 	Strikethrough bool   `json:"strikethrough,omitempty"`
